Begin to factor out ScaleContextConverter
